+++ src/Commands.cs
@@ -19,7 +19,15 @@ public class Commands
         ETGModConsole.Commands.AddGroup("hh", delegate (string[] args)
         {
             // GameManager.Instance.PrimaryPlayer.HealthAndArmorSwapped ^= true;
-            IncredibleItems.SpawnPaperChest();
+            // IncredibleItems.SpawnPaperChest();
+
+            AIActor enemyPrefab = ChewToy.ShmuppyEnemyPrefab;
+            IntVector2? randomAvailableCell = GameManager.Instance.PrimaryPlayer.CurrentRoom.GetRandomAvailableCell(new IntVector2?(enemyPrefab.Clearance), new Dungeonator.CellTypes?(enemyPrefab.PathableTiles), false);
+            if (randomAvailableCell.HasValue)
+            {
+                AIActor aIActor = AIActor.Spawn(enemyPrefab, randomAvailableCell.Value, GameManager.Instance.PrimaryPlayer.CurrentRoom, true, AIActor.AwakenAnimationType.Default, true);
+                aIActor.HandleReinforcementFallIntoRoom(0);
+            }
         });
         ETGModConsole.Commands.AddGroup("gg", delegate (string[] args)
         {