+++ src/Cwaff-Passives/ChewToy.cs
@@ -1,5 +1,6 @@
 namespace CwaffingTheGungy;
 
+using System;
 using static ShmuppyCompanion.ShmuppyMovementBehavior.State;
 
 public class ChewToy : CwaffCompanion
@@ -10,6 +11,10 @@ public class ChewToy : CwaffCompanion
     public static string Lore             = "While normally completely loyal to the Bullet Kin, Shmuppies have an incredibly soft spot for their favorite chew toys, and with minimal effort, can be coaxed into joining any friendly Gungeoneer on their adventure. They seem to be a lot more easily distracted while by your side, but their heart is in the right place.";
 
     public static string CompanionName    = "Shmuppy";
+    public static string EnemyDesc        = "Bullet Dog";
+    public static string EnemyLongDesc    = "A valued companion in bullet society. While docile and physically harmless most of the time, their barks can disrupt the passage of fired bullets and thus are a helpful way to mildly disarm trespassing gungeoneers, hopefully long enough to be stopped by reinforcements.";
+
+    public static AIActor ShmuppyEnemyPrefab = null;
 
     private static readonly string[] Directions = ["back", "back_right", "right", "front_right", "front", "front_left", "left", "back_left"];
 
@@ -18,21 +23,78 @@ public class ChewToy : CwaffCompanion
         PassiveItem item  = Lazy.SetupPassive<ChewToy>(ItemName, ShortDescription, LongDescription, Lore);
         item.quality      = ItemQuality.D;
 
-        ShmuppyCompanion friend = item.InitCompanion<ShmuppyCompanion>(friendName: CompanionName.ToID(), baseFps: 20,
+        ShmuppyCompanion friend = item.InitCompanion<ShmuppyCompanion>(friendName: CompanionName, baseFps: 20,
             extraAnims: ["bark", "hurt", "death"/*, "pitfall"*/])
           .SetPettingOffsets(new Vector2(-0.875f, -0.25f), new Vector2(0.6875f, -0.25f));
+        friend.gameObject.GetComponent<BehaviorSpeculator>().MovementBehaviors.Add(new ShmuppyCompanion.ShmuppyMovementBehavior());
 
-        tk2dSpriteAnimation library = friend.gameObject.GetComponent<tk2dSpriteAnimation>();
-        foreach (string s in Directions)
+        // init enemy variant
+        ShmuppyEnemyPrefab = CompanionName.InitEnemy(health: 30, baseFps: 20, extraAnims: ["bark", "hurt", "death"/*, "pitfall"*/],
+          shortDesc: EnemyDesc, longDesc: EnemyLongDesc);
+        ShmuppyEnemyPrefab.gameObject.GetComponent<BehaviorSpeculator>().MovementBehaviors.Add(new ShmuppyCompanion.ShmuppyMovementBehavior());
+
+        // init common stuff
+        foreach (AIActor actor in (AIActor[])[friend.aiActor, ShmuppyEnemyPrefab])
         {
-            tk2dSpriteAnimationClip idleClip = library.GetClipByName("shmuppy_idle_"+s);
-            idleClip.fps = 8;
-            tk2dSpriteAnimationClip barkClip = library.GetClipByName("shmuppy_bark_"+s);
-            barkClip.frames[2].AddSound("shmuppy_bark_sound");
+          if (actor.gameObject.GetComponent<KnockbackDoer>() is KnockbackDoer kbd)
+            kbd.weight = 30f;
+          tk2dSpriteAnimation anims = actor.gameObject.GetComponent<tk2dSpriteAnimation>();
+          foreach (string s in Directions)
+          {
+              anims.GetClipByName("shmuppy_idle_"+s).fps = 8;
+              anims.GetClipByName("shmuppy_bark_"+s).frames[2].AddSound("shmuppy_bark_sound");
+          }
         }
+    }
 
-        BehaviorSpeculator bs = friend.gameObject.GetComponent<BehaviorSpeculator>();
-        bs.MovementBehaviors.Add(new ShmuppyCompanion.ShmuppyMovementBehavior());
+    [HarmonyPatch]
+    private static class ShmuppyReinforcementPatch
+    {
+      private const float _SPAWN_CHANCE = 0.02f;
+      private static bool _SpawnedThisRoom = false;
+      private static RoomHandler _LastSpawnedRoom = null;
+
+      [HarmonyPatch(typeof(AIActor), nameof(AIActor.OnEngaged))]
+      [HarmonyPrefix]
+      private static void AIActorOnEngagedPrefixPatch(AIActor __instance, bool isReinforcement, ref bool __state)
+      {
+          __state = false; // by default, do nothing
+          if (!__instance || __instance.m_hasBeenEngaged || __instance.IsInReinforcementLayer)
+            return;
+          if (__instance.EnemyGuid != Enemies.BulletKin)
+            return;
+          // #if !DEBUG
+          if (UnityEngine.Random.value > _SPAWN_CHANCE)
+            return;
+          // #endif
+          __state = true;
+      }
+
+      [HarmonyPatch(typeof(AIActor), nameof(AIActor.OnEngaged))]
+      [HarmonyPostfix]
+      private static void AIActorOnEngagedPostfixPatch(AIActor __instance, bool isReinforcement, ref bool __state)
+      {
+        if (!__state)
+          return; // failed prefix patch checks
+        if (!__instance || __instance.CenterPosition.GetAbsoluteRoom() is not RoomHandler room || room == null)
+          return;
+        if (_LastSpawnedRoom != room)
+        {
+          _LastSpawnedRoom = room;
+          _SpawnedThisRoom = false;
+        }
+        if (_SpawnedThisRoom)
+          return;
+
+        AIActor.Spawn(
+            prefabActor     : ShmuppyEnemyPrefab,
+            position        : ShmuppyEnemyPrefab.RandomCellForEnemySpawn(room) ?? __instance.CenterPosition.ToIntVector2(),
+            source          : room,
+            correctForWalls : true,
+            awakenAnimType  : AIActor.AwakenAnimationType.Spawn)
+          .SpawnInInstantly(isReinforcement: false);
+        _SpawnedThisRoom = true;
+      }
     }
 }
 
@@ -42,10 +104,10 @@ public class ShmuppyCompanion : CwaffCompanionController
     {
         internal enum State {
             IDLE,     // doing nothing
-            FOLLOW,   // following the player
-            CHASE,    // chasing an enemy
-            BARK,     // barking at an enemy to gain their attention
-            DISTRACT, // running around an enemy once they're paying attention
+            FOLLOW,   // following its owner
+            CHASE,    // chasing its target
+            BARK,     // barking at its target to gain their attention
+            DISTRACT, // running around its target once they're paying attention
             PLAYDEAD, // playing dead during a boss fight
         }
 
@@ -54,6 +116,8 @@ public class ShmuppyCompanion : CwaffCompanionController
         private const float FAR_FROM_OWNER = 6.0f;
 
         private State _state = IDLE;
+        private HealthHaver _hh = null;
+        private UltraFortunesFavor _uff = null;
 
         public override void Start()
         {
@@ -61,6 +125,25 @@ public class ShmuppyCompanion : CwaffCompanionController
             this.retargetOnPathingFailure = true;
             m_aiActor.specRigidbody.CollideWithOthers = true; //NOTE: this doesn't work in CwaffCompanion.InitCompanion for some reason...investigate later
             m_aiActor.MovementSpeed = _SPEED;
+            this._hh = m_aiActor.gameObject.GetComponent<HealthHaver>();
+            if (!_isCompanion)
+            {
+              this._uff = m_aiActor.gameObject.AddComponent<UltraFortunesFavor>();
+              this._uff.sparkOctantVFX      = CwaffCompanionAndEnemyBuilder.FortunesFavorVFX;
+              this._uff.vfxOffset           = 0.625f;
+              this._uff.bulletRadius        = 2f;
+              this._uff.bulletSpeedModifier = 0.8f;
+              this._uff.beamRadius          = 2f;
+              this._uff.goopRadius          = 2f;
+              this._uff.enabled             = false;
+              m_aiActor.specRigidbody.OnPreRigidbodyCollision += this.OnPreRigidbodyCollision;
+            }
+        }
+
+        private void OnPreRigidbodyCollision(SpeculativeRigidbody myRigidbody, PixelCollider myPixelCollider, SpeculativeRigidbody otherRigidbody, PixelCollider otherPixelCollider)
+        {
+          if (otherRigidbody.gameObject.GetComponent<PlayerController>())
+            PhysicsEngine.SkipCollision = true;
         }
 
         private void ResetState()
@@ -72,16 +155,32 @@ public class ShmuppyCompanion : CwaffCompanionController
               enemy.OverrideTarget = null;
         }
 
+        private AIActor FindNearbyBulletKin()
+        {
+            foreach(AIActor enemy in Lazy.GetAllNearbyEnemies(m_aiActor.CenterPosition, 16f, ignoreWalls: true))
+              if (enemy.healthHaver is HealthHaver hh && hh.IsAlive && enemy.AmmonomiconName().Contains("Bullet Kin"))
+                return enemy;
+            return null;
+        }
+
         private void FollowOwner()
         {
-            this._targetActor = m_companionController.m_owner;
-            this._targetPos = this._targetActor.CenterPosition;
+            this._targetActor = _isCompanion ? m_companionController.m_owner : FindNearbyBulletKin();
+            this._targetPos = this._targetActor ? this._targetActor.CenterPosition : m_aiActor.CenterPosition;
             this._state = NearTargetActor(FAR_FROM_OWNER) ? IDLE : FOLLOW;
         }
 
         protected override void TickMovement(ref Vector2 voluntaryVel, ref Vector2 involuntaryVel)
         {
             base.TickMovement(ref voluntaryVel, ref involuntaryVel);
+            if (this._state != BARK && this._uff)
+              this._uff.enabled = false;
+            else if (this._state == BARK && this._targetActor)
+            {
+              float targetDir = (this._targetActor.CenterPosition - m_aiActor.CenterPosition).ToAngle();
+              base.m_aiAnimator.FacingDirection = targetDir;
+            }
+            // m_aiActor.DebugNametag($"{this._state.ToString()}\n{this.m_stateTimer}\n{base.m_aiAnimator.spriteAnimator.CurrentClip.name}");
         }
 
         protected override bool IsTargetValid()
@@ -98,7 +197,7 @@ public class ShmuppyCompanion : CwaffCompanionController
             return false;
         }
 
-        protected override void DetermineNewTarget()
+        private void DetermineNewTargetAsCompanion()
         {
             this._allowPathing = true;
             ResetState();
@@ -116,7 +215,28 @@ public class ShmuppyCompanion : CwaffCompanionController
             this._targetActor = enemy;
             this._targetPos = this._targetActor.CenterPosition;
             this._state = CHASE;
-            // Lazy.DebugConsoleLog($"switch to chase");
+        }
+
+        private void DetermineNewTargetAsEnemy()
+        {
+            this._allowPathing = true;
+            ResetState();
+
+            PlayerController targetPlayer = GameManager.Instance.GetRandomActivePlayer();
+            if (!targetPlayer)
+              return;
+
+            this._targetActor = targetPlayer;
+            this._targetPos = this._targetActor.CenterPosition;
+            this._state = CHASE;
+        }
+
+        protected override void DetermineNewTarget()
+        {
+            if (_isCompanion)
+              DetermineNewTargetAsCompanion();
+            else
+              DetermineNewTargetAsEnemy();
         }
 
         protected override void UpdateStateAndTargetPosition()
@@ -127,28 +247,31 @@ public class ShmuppyCompanion : CwaffCompanionController
 
         protected override bool ReachedTarget()
         {
-            if (m_companionController.IsBeingPet)
+            if (_isCompanion && m_companionController && m_companionController.IsBeingPet)
               return false;
 
             switch(this._state)
             {
-                case IDLE:     return true;                        // idle state is always reached
-                case FOLLOW:   return NearTargetActor(NEAR_OWNER); // near player
-                case CHASE:    return NearTargetActor(5f);         // near enemy
-                case BARK:     return NearTargetActor(3f);         // near enemy
-                case DISTRACT: return NearTargetPos(3f);           // near target position
-                case PLAYDEAD: return !InCombat();                 // boss battle is over
+                case IDLE:     return true;                                  // idle state is always reached
+                case FOLLOW:   return NearTargetActor(NEAR_OWNER);           // near target
+                case CHASE:    return NearTargetActor(5f);                   // near target
+                case BARK:     return NearTargetActor(3f) || StateExpired(); // near target or timeout
+                case DISTRACT: return NearTargetPos(3f) || StateExpired();   // near target or timeout
+                case PLAYDEAD: return !InCombat();                           // boss battle is over
             }
             return false;
         }
 
         private Vector2 GetRandomVisiblePointInRoom()
         {
-            if (m_companionController.m_owner is not PlayerController player)
-              return m_aiActor.CenterPosition;
-            if (player.CurrentRoom is not RoomHandler room)
-              return m_aiActor.CenterPosition;
-            return room.GetRandomVisibleClearSpot(2, 2).ToVector2();
+            RoomHandler room;
+            if (_isCompanion && m_companionController.m_owner is PlayerController player)
+              room = player.CurrentRoom;
+            else
+              room = m_aiActor.CenterPosition.GetAbsoluteRoom();
+            if (room != null)
+              return room.GetRandomVisibleClearSpot(2, 2).ToVector2();
+            return m_aiActor.CenterPosition;
         }
 
         protected override void OnReachedTarget()
@@ -176,12 +299,19 @@ public class ShmuppyCompanion : CwaffCompanionController
                         base.m_aiAnimator.PlayUntilCancelled("death");
                         break;
                       }
-                  // Lazy.DebugConsoleLog($"switch to bark");
                   this._state = BARK;
-                  base.m_aiAnimator.PlayUntilCancelled("bark");
+                  if (!this._isCompanion)
+                    this._uff.enabled = true;
+                  this.m_stateTimer = 3.0f;
                   if (this._targetActor is AIActor enemy && enemy.OverrideTarget == null)
                     enemy.OverrideTarget = this.m_aiActor.specRigidbody;
-                  // Lazy.DebugConsoleLog($"  targeting {this._targetPos.x},{this._targetPos.y}");
+                  if (this._targetActor)
+                  {
+                    float targetDir = (this._targetActor.CenterPosition - m_aiActor.CenterPosition).ToAngle();
+                    base.m_aiAnimator.FacingDirection = targetDir;
+                  }
+                  base.m_aiAnimator.EndAnimation();
+                  base.m_aiAnimator.PlayUntilCancelled("bark");
                   break;
                 }
                 case BARK:
@@ -193,20 +323,17 @@ public class ShmuppyCompanion : CwaffCompanionController
                   this._state = DISTRACT;
                   this.m_stateTimer = 5.0f;
                   this._targetPos = GetRandomVisiblePointInRoom();
-                  // Lazy.DebugConsoleLog($"switch to distract");
                   break;
                 }
                 case DISTRACT:
                 {
                   if (this.m_stateTimer <= 0.0f)
                   {
-                    // Lazy.DebugConsoleLog($"switch to follow");
                     ResetState();
                     FollowOwner();
                     break;
                   }
                   this._targetPos = GetRandomVisiblePointInRoom();
-                  // Lazy.DebugConsoleLog($"  targeting {this._targetPos.x},{this._targetPos.y}");
                   break;
                 }
                 case PLAYDEAD: