+++ changelog.md
@@ -17,6 +17,8 @@
 	- Old behavior: size (7/8/9) puzzles on (early/midgame/late) floors, random room reward chest (random quality, usually lower)
 	- New behavior: size (7-9/8-10/9-11) puzzles on (early/midgame/late) floors, chest quality scales with puzzle size
 - Added proper custom ammo clip sprites for each of Breegull's egg types
+- Added Shmuppy enemy
+	- Shmuppy sprites, portrait, and description thanks to [TitaniumGrunt7](https://www.deviantart.com/titaniumgrunt7/art/Enter-the-Gungeon-Shmuppy-715527348)
 
 ## 1.29.2 (2026-04-27)